% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbs-Rainflow_CheckRatio.R
\name{CheckRatio}
\alias{CheckRatio}
\title{A helper function for \link[rainflow]{FindPeaks}}
\usage{
CheckRatio(ai, R, aMin = NULL, aMax = NULL)
}
\arguments{
\item{ai}{Numeric, value at the current location in the time series.}

\item{R}{Numeric, the compression parameter (greater than 1).}

\item{aMin}{Numeric, value of the most recent local minimum in the time series.}

\item{aMax}{Numeric, value of the most recent local maximum in the time series.}
}
\value{
Logical indicating whether the criteria (\code{ratio > R}) has been met.
}
\description{
Checks whether a local minimum or maximum mets the criteria for 
an important local extrema based on the compression parameter \code{R}.
}
\examples{
CheckRatio(ai = 6.1, R = 1.2, aMin = 5)   # returns TRUE
CheckRatio(ai = -1.3, R = 1.1, aMax = -1) # returns TRUE
}
\author{
Addison Klinke, \email{agk38@case.edu}
}
