% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbs-Rainflow_CountCycles.numeric.R
\name{CountCycles.numeric}
\alias{CountCycles.numeric}
\title{\link[rainflow]{CountCycles} method for S3 object class \code{numeric}}
\usage{
\method{CountCycles}{numeric}(x)
}
\arguments{
\item{x}{Vector of local extrema.}
}
\value{
Object of type \code{rainflow}: a data frame with the number of cycles 
for each amplitude/mean combination.
}
\description{
Implements ASTM E1049-85 rainflow cycle counting to determine 
the amplitude and mean of each cycle within a vector (or time series) of 
alternating min/max local extrema.
}
\details{
The vector (or time series) \code{x} should strictly alternate 
min/max. If it does not, \code{CountCycles()} will automatically remove 
any repeated min or max values and issue a warning.
}
\references{
\url{https://www.astm.org/Standards/E1049.htm}
}
\seealso{
\link[rainflow]{FindPeaks}
}
\author{
Addison Klinke, \email{agk38@case.edu}
}
