% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbs-Rainflow_CountCycles.peaks.R
\name{CountCycles.peaks}
\alias{CountCycles.peaks}
\title{\link[rainflow]{CountCycles} method for S3 object class \code{peaks}}
\usage{
\method{CountCycles}{peaks}(x)
}
\arguments{
\item{p}{Object of class \code{peaks} returned from \link[rainflow]{FindPeaks}.}
}
\value{
Object of type \code{rainflow}: a data frame with the number of cycles 
for each amplitude/mean combination.
}
\description{
Implements ASTM E1049-85 rainflow cycle counting to determine 
the amplitude and mean of each cycle within a vector (or time series) of 
alternating min/max local extrema.
}
\references{
\url{https://www.astm.org/Standards/E1049.htm}
}
\seealso{
\link[rainflow]{FindPeaks}
}
\author{
Addison Klinke, \email{agk38@case.edu}
}
