% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbs-Rainflow_CountCycles.R
\name{CountCycles}
\alias{CountCycles}
\title{Generic ASTM Rainflow Cycle Counting}
\usage{
CountCycles(x)
}
\arguments{
\item{x}{Object of class \code{peaks} (preferred for best results) or 
\code{numeric}.}
}
\value{
Object of type \code{rainflow}: a data frame with the number of cycles 
for each amplitude/mean combination.
}
\description{
Implements ASTM E1049-85 rainflow cycle counting to determine 
the amplitude and mean of each cycle within a vector (or time series) of 
alternating min/max local extrema.
}
\details{
If \code{x} is class \code{numeric}, the vector should strictly 
alternate min/max. If it does not, \code{CountCycles()} will automatically 
remove any repeated min or max values and issue a warning.
}
\examples{
# For class 'numeric' (example from ASTM E1049-85 Fig. 6)
astm <- c(-2, 1, -3, 5, -1, 3, -4, 4, -2) 
r <- CountCycles(astm)
summary(r)
plot(r)

# For class 'peaks' (using package weather data)
data(weather)
p <- FindPeaks(weather$temp, R = 1.1, smooth = TRUE, window = 120)
r <- CountCycles(p)
summary(r)
plot(r)
}
\references{
\url{https://www.astm.org/Standards/E1049.htm}

\insertRef{dowling_mechanical_2013}{rainflow}
}
\seealso{
\link[rainflow]{FindPeaks}, \link[rainflow]{CountCycles.numeric},
\link[rainflow]{CountCycles.peaks}
}
\author{
Addison Klinke, \email{agk38@case.edu}
}
