% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbs-Rainflow_FindPeaks.R
\name{FindPeaks}
\alias{FindPeaks}
\title{Detect Local Extrema}
\usage{
FindPeaks(a, R, smooth, window = NULL, thres = NULL, na.pad = FALSE)
}
\arguments{
\item{a}{Vector of measured values (can be time series or not).}

\item{R}{Numeric value greater than 1 indicating the threshold for important 
local extrema. Larger values mean fewer peaks are selected (higher compression 
ratio). Try starting with 1.1 if you are unsure what value to use.}

\item{smooth}{Logical, should a moving average filter be applied before extracting 
local extrema? If \code{TRUE} (recommended), this is implemented with the 
\code{rollmean()} function from the \code{zoo} package.}

\item{window}{Numeric, if \code{smooth = TRUE}, a value indicating the window of 
points to be used for moving average filter. Else if \code{smooth = FALSE}, this 
can be left as \code{NULL} (the default).}

\item{thres}{Numeric value indicating that points in \code{a} below \code{thres}
will be dropped. This occurs before any smoothing or detection of local extrema. If
\code{thres = NULL} (the default), all points will be retained for analysis.}

\item{na.pad}{Logical, should the compressed (\code{x, y}) vectors be padded with 
\code{NA} at the ends to make them the same length as the original data? 
Useful if you are trying to add a column to a dataframe and need to match the number 
of existing rows. Then when plotting, you can use \code{na.rm = TRUE} to plot only 
the extrema.}
}
\value{
Object of class \code{peaks}, a list containing: 
\item{indices}{Locations of detected peaks.}
\item{peaks}{y-values for detected peaks.}
\item{data}{Original data passed to this function.}
\item{smooth}{If \code{smooth = TRUE}, the smoothed data, else \code{NA}.} 
\item{window}{If \code{smooth = TRUE}, the smoothing window used, else \code{NA}.}
\item{compr}{Compression ratio (1 minus number of peaks detected / length of original vector).}
\item{corr}{Pearson correlation coefficient between original and compressed vectors.}
\item{diff}{Mean difference between original and compressed vectors.}
}
\description{
Compress a vector (or time series) by extracting only the 
important local extrema (see references for algorithm implemented).
}
\examples{
# Smooth with moving average filter of 120 minutes (2 hours)
# Extract local extrema for temperature time series with R=1.1 threshold
data(weather)
p <- FindPeaks(weather$temp[1:7000], R = 1.1, smooth = TRUE, window = 120)
summary(p)

# Plot the compressed and original time series
plot(p)
}
\references{
\insertRef{fink_indexing_2004}{rainflow}
}
\author{
Addison Klinke, \email{agk38@case.edu}
}
