% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbs-Rainflow_plot.peaks.R
\name{plot.peaks}
\alias{plot.peaks}
\title{Visualize extracted local extrema}
\usage{
\method{plot}{peaks}(p)
}
\arguments{
\item{p}{An object of type \code{peaks} returned from 
\link[rainflow]{FindPeaks}.}
}
\value{
None.
}
\description{
Given the \code{peaks} object returned from 
\link[rainflow]{FindPeaks}, \code{plot.peaks()} displays the 
original data and locations of the detected local extrema.
}
\details{
Plot shows the raw data in black, moving average smoothed data (if requested)
in green, and detected peaks in red.
}
\author{
Addison Klinke, \email{agk38@case.edu}
}
