% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbs-Rainflow_plot.rainflow.R
\name{plot.rainflow}
\alias{plot.rainflow}
\title{Visualize rainflow cycles}
\usage{
\method{plot}{rainflow}(r)
}
\arguments{
\item{r}{An object of type \code{rainflow} returned from 
\link[rainflow]{CountCycles}.}
}
\value{
None.
}
\description{
Given the \code{rainflow} object returned from 
\link[rainflow]{CountCycles}, \code{plot.rainflow()} displays the 
a scatter plot of cycle mean and amplitude.
}
\author{
Addison Klinke, \email{agk38@case.edu}
}
