% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbs-Rainflow_summary.peaks.R
\name{summary.peaks}
\alias{summary.peaks}
\title{S3 method for object of type \code{peaks}}
\usage{
\method{summary}{peaks}(p)
}
\arguments{
\item{p}{An object of type \code{peaks} returned from 
\link[rainflow]{FindPeaks}.}
}
\value{
None.
}
\description{
S3 method for object of type \code{peaks}
}
\author{
Addison Klinke, \email{agk38@case.edu}
}
