% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbs-Rainflow_summary.rainflow.R
\name{summary.rainflow}
\alias{summary.rainflow}
\title{S3 method for object of type \code{rainflow}}
\usage{
\method{summary}{rainflow}(r)
}
\arguments{
\item{r}{An object of type \code{rainflow} returned from 
\link[rainflow]{CountCycles}.}
}
\value{
None.
}
\description{
S3 method for object of type \code{rainflow}
}
\author{
Addison Klinke, \email{agk38@case.edu}
}
